Валидация пароля вынесена в компонен статусбар. Исправлены некоторые элементы
