+++ src/api/api.js
@@ -58,7 +58,7 @@ export const api = {
 
         return data;
       } catch (e) {
-        console.log(e);
+        console.log(e)
       }
     },
   },
